[office] Fallback to an offline install of MS365
When the online installation via
https://www.microsoft365.com/from=Homescreen fails, attempt an offline
installation by hardcoding the manifest from
https://www.microsoft365.com/webmanifest.json.

Add the new histogram
FileBrowser.OfficeFiles.Setup.OfficeWebAppOfflineInstallation to record
the result.

Save the MS365 icon to be able to display it when it can't be
downloaded.

Also move the title of the MS365 app (IDS_OFFICE_FILE_HANDLER_APP_MICROSOFT) from chromeos_strings.grdp to
chromeos_shared_strings.grdp so it can be used when in Lacros.

Example:
https://screencast.googleplex.com/cast/NDkxMjgzOTEzMDA4NzQyNHw4NWI1Njk5MS01YQ

Bug: b/340978744
Change-Id: If175bca79d3d40b1604d007a8c1ce175a2bdf7a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5599980
Reviewed-by: Alexander Bolodurin <alexbn@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Cassy Chun-Crogan <cassycc@google.com>
Cr-Commit-Position: refs/heads/main@{#1311740}