launcher: Always record image info into the database when processed.
There is a bug that when an image has no valid annotation, it's never
get recorded and will be re-processed whenever the user logs in.
This CL fixes it by always inserting image info into db even if the
annotation list is empty.

More details in the bug.

Bug: b/345407818
Change-Id: If54b201d8ee781cebe574a584ca4b50e84d9dd51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5602632
Reviewed-by: Charles Zhao <charleszhao@chromium.org>
Commit-Queue: CJ Huang <chenjih@google.com>
Cr-Commit-Position: refs/heads/main@{#1311757}