focus_mode: Implement a functional skeleton of the web audio player
This CL adds the audio playing part of the focus mode web ui. It adds a
new page (chrome-untrusted://focus-mode-player) that is loaded as an
iframe in chrome://focus-mode-media. The trusted outer frame
communicates with the browser via a mojom interface.

The inner page is needed because pages that communicate with the browser
(like the outer one does via mojom) are not allowed to make external
network requests.

The controller is currently not hooked up to the rest of the focus mode
and will simply respond to requests with a hardcoded track.

BUG=b:337064323

Change-Id: I16ff29f95484e5815bf4d563d153709df09b7411
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5560397
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311765}