Add new Webstore extension installation status for supervised users
This change introduces a new installation state,
`kCustodianApprovalRequiredForInstallation`, which should be used
for new extension installations that require parental approval.
It is only returned if the corresponding feature is enabled, until
the installation flow for new extensions for supervised users is
adjusted in chromium and the webstore server side.

The existing `kCustodianApprovalRequired` state, may still be used
for the case than an installed extension is missing parental approval.
If this is consumed by them Webstore to display the
Supervised User banner, then we need to ensure that the action it
triggers is "Set extension to enabled" (and not "Install extension"
which is what happens today).

Bug: b/333713552
Change-Id: I42739cd2d5e2e4f4caa4a297eeb68fec4c5d77fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5573785
Commit-Queue: Tomek Jurkiewicz <tju@google.com>
Auto-Submit: Anthi Orfanou <anthie@google.com>
Reviewed-by: Tomek Jurkiewicz <tju@google.com>
Reviewed-by: Tim <tjudkins@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311862}