AW: Add annotation to run tests in only single or multi process mode
No change to production logic, only tests.

This adds a new annotation @OnlyRunIn(EITHER_PROCESS) which runs tests
in only one mode, rather than duplicating the test in both modes (which
is currently the default). The new enum currently defaults to only
running tests in multi-process mode as this is used by the majority of
devices in production.

The motivation to have this separate enum is to be able to differentiate
between tests that are only being run in a single mode for efficiency or
that are only being run in a single mode because they cannot run in
both. This will allow us to be able to control which tests run in
different test environments and on which builders, e.g. tests labelled
@OnlyRunIn(SINGLE_PROCESS) might be skipped in CQ on O+ devices. This
will reduce costs of running tests on chromium infra without losing test
coverage from the tests labelled as @OnlyRunIn(SINGLE_PROCESS) because
they only need to run in one mode for efficiency.

This change only updates tests explicitly labelled/commented as unit
tests/tests that only need to in one mode to use the
@OnlyRunIn(EITHER_PROCESS) annotation. A follow up CL will update other
files with specific cases that could use this annotation.

Bug: 341913802
Change-Id: Ie564152c7ed190a5eb0ec4064b2aaf117f7dbe2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5593829
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Apoorva Verma <apoorvaverma@google.com>
Cr-Commit-Position: refs/heads/main@{#1311880}