CBD Timeframe Experiment: error announcement via focus on dropdown
When users click the 'Clear' button without having selected a time
period, then the requirement to select a time period should be read
out to screen reader users. This CL removes the explicit announcement
for this, as this is already achieved by the focus being moved to the
dropdown, which explains this requirement through the text of its
current selection.

Bug: 40283307
Change-Id: I688b1c90d374175063ac5fbb20543eaf25f786e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5600855
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Commit-Queue: Rainhard Findling <rainhard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311885}