[Passwords, hats] Trigger Password Manager happiness survey.
Start triggering a user survey when the user fills a form manually
and submits it (successfully or unsuccessfully). The survey is
triggered at the moment when we are sure the submission has happened.
Only manual fillings are included, forms filled on pageload are
excluded, to filter users who are more likely to be aware they've used
a Password Manager suggestion.

Bug: 338341118
Change-Id: I5b43906bf745daa7c70ebb15e17d512200545c94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5604224
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/main@{#1311892}