Fix easing fn not being applied at the bounds of an animation
The ...AnimationCurve:GetValue methods assume that the boundary keyframe
values are used at the endpoints of the animation. This is not correct
if using a step timing function with a discontinuity at the start or
end. In all cases, the timing function should be applied, though the time should be clamped to the range.

Bug: 333533025
Change-Id: If6a1534c90b8ede57e8adb58a50e2a6677e5df29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5581881
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311907}