Add metric to record ScreenAI requests' success.
To monitor success of ScreenAI service, metrics are needed that record
whether the requests to OCR and Main Content Extraction were executed
successfully or failed. This is regardless of whether the reply
contained any data or not as that is sensitive to the inputs to the
library.

AX-Relnotes: n/a
Bug: 342796806
Change-Id: I190a88a83688d5153f0ff5f9cf51016bae0f0fe3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5606423
Auto-Submit: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1311950}