Add UMA metrics for plus address submission logging.
On top of that, this CL introduces a new directory for plus address
metrics. The existing plus address metrics are part of the autofill
histograms file, but as //components/plus_addresses continues to grow,
this is suboptimal. Existing metrics will eventually be migrated here,
too.

Bug: 343124027
Change-Id: I62f5f3320fb6455f11be84d41dcd4ffb228fdc27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5594224
Reviewed-by: Christoph Schwering <schwering@google.com>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Matt Reichhoff <mreichhoff@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311957}