Test migration of proxied connections on network change
The connection to the destination should not migrate, as it is carried
by the connection to the proxy. The connection to the proxy, however,
should migrate like any other connection.

Luckily, this was already the case: proxied sessions have
`kInvalidNetworkHandle`, which never matches the handle in a network
change notification.

However, testing uncovered a UAF bug, where a `handles::NetworkHandle*` pointing to a value on the stack was passed to a callback that ran after that stack frame was deallocated. This CL threads a pointer to the session attempt's `network_` through a few methods to avoid this issue.

Bug: 338342162
Change-Id: I53e6a63247b4c3824aacc4775bf875f4936feaf5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5593530
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Fan Yang <fayang@chromium.org>
Commit-Queue: Dustin Mitchell <djmitche@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311995}