Sql: change SQL script parameter type from char* to cstring_view
And some other related updates.

GetUniqueStatement is updated to take a base::cstring_view, which is a
null-terminated string that can be either a char[] or std::string.

GetCachedStatement should very rarely be dynamically constructed, as
it should almost always be statically known. Thus it has less need to
take a `std::string`. However, taking a `char*` has not prevented
callers from using a `string` with `c_str()`, whether or not the
string was necessary or appropriate. Therefore we might as well
update GetCachedStatement as well for consistency.

Bug: 341639215

Change-Id: If034a0db0f6546ac54dd99dbc423f1a76bf02727
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5571204
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Elad Alon <eladalon@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Nathan Memmott <memmott@chromium.org>
Reviewed-by: Angela Yoeurng <yoangela@chromium.org>
Reviewed-by: Steven Bingler <bingler@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Matt Reichhoff <mreichhoff@chromium.org>
Reviewed-by: Cassy Chun-Crogan <cassycc@google.com>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Austin Sullivan <asully@chromium.org>
Reviewed-by: Max Curran <curranmax@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Reviewed-by: Patrick Meenan <pmeenan@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1312004}