[Views AX] Lazily initialize product strings
When investigating the fuzzer bugs that started hitting after
CL:5493594 landed, I started wondering if retrieving this
information really should be happening at startup. These are
tied to a11y, and are really exposed on windows via IA2.

With this in mind, this change moves retrieval to when the data
is needed, while still caching the result to avoid constantly
hopping to the content client whenever this is called.

This change also removed a redundant CHECK that was introduced
in the prior change to improve readability.
We already have a CHECK_DEREF, so the check above it was no
longer needed.

Finally, this change also moves this logic to be windows-only,
as these are only needed for IA2 support.

Bug: 341060174, 340643919, 341060174
Change-Id: Ic08f5b281dfeed78cdeebeea819c55abd5441697
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5546286
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Jacques Newman <janewman@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1312019}