Batched Main-thread Resource Release
Context switching to the Main-thread can be expensive. For
ClientResourceProvider::ReceiveReturnsFromParent we showde that baching
all of the Main-thread release callbacks was beneficial.

This adds support to the other resource removing paths.

This creates ClientResourceProvider::ScopedBatchResourcesRelease to
denote when a series of separate events can all result in the removal
of resources. So that we can batch across them. Such as when a new tree
is being committed.

Bug: 40269731, 40250049
Change-Id: Ic2118ae06def86474bead0b88840299b36a5af08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5573827
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1312023}