[SBER Deprecation] Remove unused extended-reporting-related code
ExtendedReportingPrefExists function is read in security interstitial
and download protection code. In some cases it is called before
checking the value of the preference, in this case the function is not
needed because the feature has been fully rolled out for a long time so
the preference value exists. In other cases, the return value is stored
in a variable and no longer used. This CL removes this function to clean
the code base as part of the Safe Browsing Extended Reporting
deprecation.

Bug: 336547987
Change-Id: I2f47eed5c0ae6f373f46ed0a0103868d0195bbb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5587901
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Commit-Queue: Zack Han <zackhan@chromium.org>
Reviewed-by: Javier Castro <jacastro@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1312037}