[Extensions cleanup] Remove uninstall dialog strings only used in test
ExtensionUninstallDialog used to provide the strings to the dialog via
GetHeadingText(). However, this was changed a while ago but the method
was still being used by some tests. This is not correct, since it's
not checking the actual content of the dialog.

Thus, this CL removes GetHeadingText() and its usage in the test (it's
effectively not testing anything).

Bug: 342369756
Change-Id: I0cca9b6d38ece8578eb9c6214487b381739cd0ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5594772
Reviewed-by: Justin Lulejian <jlulejian@chromium.org>
Commit-Queue: Justin Lulejian <jlulejian@chromium.org>
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1312061}