[Signin] On Sign out cancel access token fetchers with persistent error
Currently on signout, access token fetchers are cancelled with
`GoogleServiceAuthError::REQUEST_CANCELLED` which is a valid transient
error. Consumers of access tokens may retry the fetch which might lead
to a crash if this happens during signout. Therefore this CL updates the
error to be `USER_NOT_SIGNED_UP` when credentials are revoked which
aligns with the error sent when there is no refresh token for the
account. This error is persistent and clients are expected not to retry.

Bug: b/321900823
Change-Id: I2f4bc54c658d3124c72b738bf3655bc1798ff0d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5563164
Reviewed-by: David Roger <droger@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1312108}