[ios] Explicitly create OverlayRequestQueue
The presentation of infobar will access the OverlayRequestQueue
which is lazily created. Due to how the destruction of WebState
happens (first it destroy the CRWWebController which causes the
JavaScript callbacks to be invoked, then it notify observers of
the destruction, then destroy the tab helper), this could lead
a OverlayRequestQueue to be created during the destruction of
the WebState and a crash.

Instead explicitly create OverlayRequestQueue like other tab
helpers when the WebState is inserted in the WebStateList (or
when it is used for Prerender) to ensure it is not created at
an inopportune time.

Bug: 345268433, 345271453
Change-Id: I551b9e17b63064c271bb469e184189534685b271
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5607673
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1312126}