Add NotifyClipboardDataChanged call to ClipboardWin
This CL is required to have `CopyObserver` work in
`ClipboardHostImpl`. This was the cause of b/343945726 as
clipboard writes not notifying this observer resulted in
DLP functionality not working.

The `NotifyClipboardDataChanged()` call has to happen right
after `CloseClipboard()` so that observers get notified immediately
after the clipboard data has changed and a new seqno is
available for it.

Bug: b/343945726
Change-Id: I945428e9ac79db60bdff744c6de432ef1d0efd38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5588381
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Dominique Fauteux-Chapleau <domfc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1312143}