Fix dangling pointer when a ProofVerifierChromium Job has live Request.
ProofVerifierChromium::Job previously had a CertVerifier::Request that
was declared above a ProofVerifyDetailsChromium. The Request is passed
a raw pointer to the CertVerifyResult, and is expected to hand onto it.
Thus, on cancellation, a dangling raw pointer warning would be thrown,
due to tearing down the Result before the Request that had a raw_ptr to
it.

This CL flips the order of the two objects, and adds a unit tests.

Fixed: 345518606
Change-Id: I7a0009fe7e7089843a35dca13ad47c995fd4ae38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5606902
Commit-Queue: mmenke <mmenke@chromium.org>
Reviewed-by: Matt Mueller <mattm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1312160}