[Tab Group Parity] On NTP open in new tab in group trigger creation dialog
We didn't realize NTP and the tab longpress menu had separate delegates
for opening new tabs in tab groups. Address this missed action.

This was overlooked in:
https://chromium-review.googlesource.com/c/chromium/src/+/5604931

I think the one test for this in NewTabPageTest is disabled and the
change itself is very minor.

Low-Coverage-Reason: TESTS_ARE_DISABLED
Fixed: 345611795
Change-Id: I234fa41b808fc54e5edf06dff2aa52afa16d6d41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5605666
Reviewed-by: Xi Han <hanxi@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1312168}