[E2E] Use BottomControlsStacker to resize bottom browser controls
Introduce BrowserControlsStacker to coordinate different UI that wants to change the bottom controls height. This CL is used as transitional solution, just forwarding calls from BrowserControlsStacker to the wrapped BrowserControlsSizer, while also provide the BrowserControlsStateProvider for clients to do the observation.

There are still references of BrowserControlsSizer in custom_tabs, which  is not in scope of this initial refactor.

Bug: 345488108
Change-Id: Ia0aceefba7909deceb0dcefe7e39a0a88aeb03d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5605415
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Wenyu Fu <wenyufu@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1312201}