Fix regression with backing up notification images.
In crrev.com/c/5405074, in the user journey where the user presses the
"Unsubscribe" button on a notification, the order of backing up
notification image resources, and replacing the notification with the
`provisionally unsubsribed` notification has been accidentally reversed.
This led to image resources not being restored correctly if the user
chooses to "Undo" the unsubscribe action.

This CL fixes this and adds a regression test.

Bug: 345640570
Change-Id: I58fc4eb1e1e8d0b066ea142290ecf443ae414b84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5607790
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1312205}