[FedCM] Add registered IDP types
This feature allows the IDP to specify `types` in its config URL. An RP
may specify a type when invoking get(). If an RP does not specify
a type, all registered IDPs are used. If an RP specifies a type, only
IdPs having a matching one in their `types` array are used.

Bug: 341145947
Change-Id: I8c0d410c06cecf8132bd55174f8a6437db9538b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5546318
Commit-Queue: Nicolás Peña <npm@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Sam Goto <goto@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1312224}