[NTP][Enterprise] Require user is managed to see Calendar card
Pulls the code Drive was using to check if the user was managed into
a separate method to pass through to get the list of modules ids.

I looked into addressing the removal of the private method, but I was
running into issues. I left it as continuing to use the call for now
with the bug ticket still attached. It is something that needs to be
addressed, but not as part of this.

Bug: 344897043
Change-Id: Ifc69efe0c934359c3b0c1fc5d65b59ecc437accc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5605419
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Riley Tatum <rtatum@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1312232}